remove duplicate code part 4
