+++ cmd/frontend/graphqlbackend/search_test.go
@@ -12,6 +12,7 @@ import (
 	"github.com/google/go-cmp/cmp/cmpopts"
 	"github.com/graph-gophers/graphql-go"
 
+	searchrepos "github.com/sourcegraph/sourcegraph/cmd/frontend/internal/search/repos"
 	"github.com/sourcegraph/sourcegraph/internal/api"
 	"github.com/sourcegraph/sourcegraph/internal/conf"
 	"github.com/sourcegraph/sourcegraph/internal/db"
@@ -329,7 +330,7 @@ func TestExactlyOneRepo(t *testing.T) {
 	}
 	for _, c := range cases {
 		t.Run("exactly one repo", func(t *testing.T) {
-			if got := exactlyOneRepo(c.repoFilters); got != c.want {
+			if got := searchrepos.ExactlyOneRepo(c.repoFilters); got != c.want {
 				t.Errorf("got %t, want %t", got, c.want)
 			}
 		})
@@ -577,104 +578,6 @@ func TestVersionContext(t *testing.T) {
 	}
 }
 
-func TestComputeExcludedRepositories(t *testing.T) {
-	cases := []struct {
-		Name              string
-		Query             string
-		Repos             []types.Repo
-		WantExcludedRepos excludedRepos
-	}{
-		{
-			Name:  "filter out forks and archived repos",
-			Query: "repo:repo",
-			Repos: []types.Repo{
-				{
-					Name: "repo-ordinary",
-				},
-				{
-					Name: "repo-forked",
-					Fork: true,
-				},
-				{
-					Name: "repo-forked-2",
-					Fork: true,
-				},
-				{
-					Name:     "repo-archived",
-					Archived: true,
-				},
-			},
-			WantExcludedRepos: excludedRepos{forks: 2, archived: 1},
-		},
-		{
-			Name:  "exact repo match does not exclude fork",
-			Query: "repo:^repo-forked$",
-			Repos: []types.Repo{
-				{
-					Name: "repo-forked",
-					Fork: true,
-				},
-			},
-			WantExcludedRepos: excludedRepos{forks: 0, archived: 0},
-		},
-		{
-			Name:  "when fork is set don't populate exclude",
-			Query: "repo:repo fork:no",
-			Repos: []types.Repo{
-				{
-					Name: "repo",
-				},
-				{
-					Name: "repo-forked",
-					Fork: true,
-				},
-			},
-			WantExcludedRepos: excludedRepos{forks: 0, archived: 0},
-		},
-	}
-
-	for _, c := range cases {
-		// Setup: parse the query, extract its repo filters, and use
-		// those to populate the resolve repo options to pass to the
-		// function under test.
-		q, err := query.ParseAndCheck(c.Query)
-		if err != nil {
-			t.Fatal(err)
-		}
-		r := searchResolver{query: q}
-		includePatterns, _ := r.query.RegexpPatterns(query.FieldRepo)
-		options := db.ReposListOptions{IncludePatterns: includePatterns}
-
-		// Setup: the mock DB lookup returns forked repo count if OnlyForks is set,
-		// and archived repo count if OnlyArchived is set.
-		db.Mocks.Repos.Count = func(_ context.Context, options db.ReposListOptions) (int, error) {
-			var count int
-			if options.OnlyForks {
-				for _, repo := range c.Repos {
-					if repo.Fork {
-						count += 1
-					}
-				}
-			}
-			if options.OnlyArchived {
-				for _, repo := range c.Repos {
-					if repo.Archived {
-						count += 1
-					}
-				}
-			}
-			return count, nil
-		}
-
-		t.Run("exclude repo", func(t *testing.T) {
-			got := computeExcludedRepositories(context.Background(), q, options)
-			if !reflect.DeepEqual(got, c.WantExcludedRepos) {
-				t.Fatalf("results = %+v, want %+v", got, c.WantExcludedRepos)
-			}
-		})
-	}
-}
-
 func mkFileMatch(repo *types.RepoName, path string, lineNumbers ...int32) *FileMatchResolver {
 	if repo == nil {
 		repo = &types.RepoName{
@@ -721,57 +624,5 @@ func TestGetReposWrongUnderlyingType(t *testing.T) {
 	_, err := getRepos(context.Background(), rp)
 	if err == nil {
 		t.Errorf("Expected error, got nil")
-	}
-}
-
-func TestHasTypeRepo(t *testing.T) {
-	tests := []struct {
-		query           string
-		wantHasTypeRepo bool
-	}{
-		{
-			query:           "sourcegraph type:repo",
-			wantHasTypeRepo: true,
-		},
-		{
-			query:           "sourcegraph type:symbol type:repo",
-			wantHasTypeRepo: true,
-		},
-		{
-			query:           "(sourcegraph type:repo) or (goreman type:repo)",
-			wantHasTypeRepo: true,
-		},
-		{
-			query:           "sourcegraph repohasfile:Dockerfile type:repo",
-			wantHasTypeRepo: true,
-		},
-		{
-			query:           "repo:sourcegraph type:repo",
-			wantHasTypeRepo: true,
-		},
-		{
-			query:           "repo:sourcegraph",
-			wantHasTypeRepo: false,
-		},
-		{
-			query:           "repository",
-			wantHasTypeRepo: false,
-		},
-		{
-			query:           "",
-			wantHasTypeRepo: false,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.query, func(t *testing.T) {
-			q, err := query.ProcessAndOr(tt.query, query.ParserOptions{SearchType: query.SearchTypeLiteral})
-			if err != nil {
-				t.Fatal(err)
-			}
-			if got := hasTypeRepo(q); got != tt.wantHasTypeRepo {
-				t.Fatalf("got %t, expected %t", got, tt.wantHasTypeRepo)
-			}
-		})
 	}
 }
